Move Windows path test function to test code
